rc5 for adding Guid in folder names
