+++ src/knmidownloader/Discord/Message.cs
@@ -1,0 +1,15 @@
+﻿using Discord;
+using Discord.Commands;
+using Discord.Net;
+using Discord.WebSocket;
+
+namespace knmidownloader.Discord
+{
+    internal class Message
+    {
+        public Message()
+        {
+            // unused still
+        }
+    }
+}
